requirements.txt
fastapi
uvicorn
torch
pillow
numpy
python-multipart
torchvision
